sort files into proper subfolders
